DSND-2930: Refactor business logic
* Rename processor to service so controller depends on service
* Service depends on a repository interface
* The repository implementation depends on a specific MongoRepository interface
  and handles mongo specific exceptions.
