--- src/main/java/uk/gov/companieshouse/acspprofile/api/service/GetProcessor.java
@@ -1,10 +1,0 @@
-package uk.gov.companieshouse.acspprofile.api.service;
-
-import uk.gov.companieshouse.api.acspprofile.AcspFullProfile;
-import uk.gov.companieshouse.api.acspprofile.AcspProfile;
-
-public interface GetProcessor {
-
-    AcspProfile getProfile(String acspNumber);
-    AcspFullProfile getFullProfile(String acspNumber);
-}
